// src/survey/items/types.ts
export enum SurveyItemType {
  Group = 'group',
  Display = 'display',
  PageBreak = 'pageBreak',
  SurveyEnd = 'surveyEnd',

  SingleChoiceQuestion = 'singleChoiceQuestion',
  MultipleChoiceQuestion = 'multipleChoiceQuestion',
}



export enum ConfidentialMode {
  Add = 'add',
  Replace = 'replace'
}